generates a basic sql statement
The basic test so far involves using the parser to get take the test
SQL files and translate them into an AST. We then generate SQL from
the AST and check the following:
* cant the output be reparsed by the original parser without error
* is the output (stripped of whitespace) the same as the original

These are very coarse tests but they allow for the original to be
heavily refactored without being too coupled.

There is currently no reason to use the ast traverser other than
for modification of the AST. The grammer is too complicated to use
the traverser visitors as the main method of building strings.