started working on the test suite
the entire suite is autogenerated from the test cases provided to it.
adding a new squite involves creating a subfolder in the sql folder
and adding an sql statement. The next time the test suite is run it
will create a describe and it block for the new statement. this does
however mean that there are 65 failing tests and currently only a
single passing test.